Feat: add code line for encoding type change
